+++ lib/parser.ml
@@ -243,7 +243,7 @@ let lbp = function
 
 let is_atom_start = function
   | Token.Int _ | Token.Float _ | Token.String _ | Token.Bool _
-  | Token.Path _ | Token.Glob _ | Token.Date _ | Token.Time _ | Token.DateTime _
+  | Token.Path _ | Token.Glob _ | Token.DateTime _
   | Token.Duration _ | Token.Url _ | Token.IPv4 _ | Token.CIDR _
   | Token.Port _ | Token.Version _ | Token.Size _
   | Token.Ident _ | Token.Upper _ | Token.Hole
@@ -287,7 +287,7 @@ let collapse_multi_equation arity eqs : Ast.pat list * Ast.expr =
 
 let builtin_types = [
   "Int"; "Float"; "String"; "Bool"; "Unit"; "Path";
-  "Date"; "Time"; "DateTime"; "Duration"; "Url";
+  "DateTime"; "Duration"; "Url";
   "IPv4"; "CIDR"; "Port"; "Version"; "Size"
 ]
 
@@ -697,8 +697,6 @@ and atom_base_ s =
   | Token.Bool b     -> Bool b
   | Token.Path p     -> Path p
   | Token.Glob g     -> Glob g
-  | Token.Date d     -> Date d
-  | Token.Time t     -> Time t
   | Token.DateTime d -> DateTime d
   | Token.Duration d -> Duration d
   | Token.Url u      -> Url u